requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.23.0
scipy>=1.10.0
pillow>=9.5.0
tqdm>=4.65.0